IntentTools: move functions from ViewModel here
