+++ app/src/main/java/page/ooooo/geoshare/lib/IntentTools.kt
@@ -1,11 +1,25 @@
 package page.ooooo.geoshare.lib
 
+import android.content.ActivityNotFoundException
 import android.content.ComponentName
+import android.provider.Settings
+import android.content.Context
 import android.content.Intent
+import android.content.pm.PackageManager
+import android.graphics.drawable.Drawable
 import android.net.Uri
+import android.os.Build
 import android.util.Log
+import android.widget.Toast
+import androidx.activity.result.ActivityResultLauncher
+import androidx.core.net.toUri
+import page.ooooo.geoshare.BuildConfig
+import page.ooooo.geoshare.R
+import java.util.Locale
 
 class IntentTools {
+
+    data class App(val packageName: String, val label: String, val icon: Drawable)
 
     fun createViewIntent(packageName: String, data: Uri): Intent = Intent(Intent.ACTION_VIEW, data).apply {
         setPackage(packageName)
@@ -47,6 +61,97 @@ class IntentTools {
                 Log.w(null, "Unsupported intent action $intentAction")
                 return null
             }
+        }
+    }
+
+    fun queryGeoUriApps(packageManager: PackageManager): List<App> {
+        val resolveInfos = try {
+            packageManager.queryIntentActivities(
+                Intent(Intent.ACTION_VIEW, "geo:".toUri()),
+                PackageManager.MATCH_DEFAULT_ONLY,
+            )
+        } catch (e: Exception) {
+            Log.e(null, "Error when querying apps that support geo: URIs", e)
+            return emptyList()
+        }
+        return resolveInfos.mapNotNull {
+            try {
+                App(
+                    it.activityInfo.packageName,
+                    it.activityInfo.loadLabel(packageManager).toString(),
+                    it.activityInfo.loadIcon(packageManager),
+                )
+            } catch (e: Exception) {
+                Log.e(null, "Error when loading info about an app that supports geo: URIs", e)
+                null
+            }
+        }.filterNot { it.packageName == BuildConfig.APPLICATION_ID }.sortedBy { it.label }
+    }
+
+    fun openApp(context: Context, packageName: String, uriString: String) {
+        try {
+            context.startActivity(createViewIntent(packageName, uriString.toUri()))
+        } catch (_: ActivityNotFoundException) {
+            Toast.makeText(context, R.string.conversion_succeeded_apps_not_found, Toast.LENGTH_SHORT).show()
+        }
+    }
+
+    fun openChooser(context: Context, uriString: String) {
+        try {
+            context.startActivity(createChooserIntent(uriString.toUri()))
+        } catch (_: ActivityNotFoundException) {
+            Toast.makeText(context, R.string.conversion_succeeded_apps_not_found, Toast.LENGTH_SHORT).show()
+        }
+    }
+
+    fun isDefaultHandlerEnabled(packageManager: PackageManager, uriString: String): Boolean {
+        val resolveInfo = try {
+            packageManager.resolveActivity(
+                Intent(Intent.ACTION_VIEW, uriString.toUri()),
+                PackageManager.MATCH_DEFAULT_ONLY,
+            )
+        } catch (e: Exception) {
+            Log.e(null, "Error when querying which app is the default handler for a URI", e)
+            return false
+        }
+        val packageName = try {
+            resolveInfo?.activityInfo?.packageName
+        } catch (e: Exception) {
+            Log.e(null, "Error when loading info about an app that is the default handler for URI", e)
+            null
+        }
+        return packageName == BuildConfig.APPLICATION_ID
+    }
+
+    fun showOpenByDefaultSettings(
+        context: Context,
+        launcher: ActivityResultLauncher<Intent>,
+    ) {
+        showOpenByDefaultSettingsForPackage(context, launcher, BuildConfig.APPLICATION_ID)
+    }
+
+    fun showOpenByDefaultSettingsForPackage(
+        context: Context,
+        launcher: ActivityResultLauncher<Intent>,
+        packageName: String,
+    ) {
+        try {
+            val action = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S &&
+                // Samsung supposedly doesn't allow going to the "Open by default" settings page.
+                Build.MANUFACTURER.lowercase(Locale.ROOT) != "samsung"
+            ) {
+                Settings.ACTION_APP_OPEN_BY_DEFAULT_SETTINGS
+            } else {
+                Settings.ACTION_APPLICATION_DETAILS_SETTINGS
+            }
+            val intent = Intent(action, "package:$packageName".toUri())
+            launcher.launch(intent)
+        } catch (_: ActivityNotFoundException) {
+            Toast.makeText(
+                context,
+                R.string.intro_settings_activity_not_found,
+                Toast.LENGTH_LONG,
+            ).show()
         }
     }
 }
